Multiple SortableTrees can be one one page. Sorry to say eslint forced me to push the SortableTree component into a different file. You guess can fix the naming of these files as seems fit.
